feat: separate injection logics
