+++ p2p_wallet/Injection/Resolver.registerAllServices.swift
@@ -1,0 +1,19 @@
+import Resolver
+
+extension Resolver: ResolverRegistering {
+    @MainActor public static func registerAllServices() {
+        // For lifetime app's services
+        registerForApplicationScope()
+
+        registerForGraphScope()
+
+        // For services that lives inside user's session
+        registerForSessionScope()
+
+        registerForSharedScope()
+    }
+}
+
+extension ResolverScope {
+    static let session = ResolverScopeCache()
+}
